+++ packages/playwright/src/index.ts
@@ -5,12 +5,11 @@ import type {
   AxeResults,
   SerialContextObject,
   PartialResults,
-  ContextObject,
   SerialSelectorList,
-  SerialSelector,
   SerialFrameSelector
 } from 'axe-core';
-import { source } from 'axe-core';
+import axe from 'axe-core';
+const { source } = axe;
 import { normalizeContext, analyzePage } from './utils';
 import type { AxePlaywrightParams } from './types';
 import {
@@ -150,7 +149,7 @@ export default class AxeBuilder {
 
   public async analyze(): Promise<AxeResults> {
     const context = normalizeContext(this.includes, this.excludes);
-    const { page, option: options } = this;
+    const { page } = this;
 
     page.evaluate(this.script());
     const runPartialDefined = await page.evaluate<boolean>(
@@ -326,3 +325,5 @@ export default class AxeBuilder {
     `;
   }
 }
+
+export { AxeBuilder };