+++ packages/puppeteer/src/utils.ts
@@ -1,7 +1,10 @@
+import assert from 'assert';
 import * as fs from 'fs';
 import * as Axe from 'axe-core';
 import { Frame } from 'puppeteer';
 import { axeConfigure, axeShadowSelect } from './browser';
+import { getFilename } from 'cross-dirname';
+import { pathToFileURL } from 'url';
 import { pageIsLoaded } from './browser';
 
 export async function frameSourceInject(
@@ -11,8 +14,22 @@ export async function frameSourceInject(
 ): Promise<void> {
   await assertFrameReady(frame);
   if (!source) {
-    const pathFile = require.resolve('axe-core');
-    source = fs.readFileSync(pathFile, 'utf8');
+    let axeCorePath = '';
+    if (
+      typeof require === 'function' &&
+      typeof require.resolve === 'function'
+    ) {
+      axeCorePath = require.resolve('axe-core');
+    } else {
+      const { createRequire } = (await import('node:module')) as any;
+      // `getFilename` is needed because esm's `import.meta.url` is illegal syntax in cjs
+      const filename = pathToFileURL(getFilename()).toString();
+
+      const require = createRequire(filename);
+      axeCorePath = require.resolve('axe-core');
+    }
+
+    source = fs.readFileSync(axeCorePath, 'utf8');
   }
   await frame.evaluate(source);
   await frame.evaluate(axeConfigure, config as Axe.Spec);
@@ -57,14 +74,18 @@ export async function getChildFrame(
 export async function assertFrameReady(frame: Frame): Promise<void> {
   // Wait so that we know there is an execution context.
   // Assume that if we have an html node we have an execution context.
-  // Check if the page is loaded.
-  let pageReady = false;
   try {
-    pageReady = await frame.evaluate(pageIsLoaded);
+    // Puppeteer freezes on unloaded iframes. Set a race timeout in order to handle that.
+    // @see https://github.com/dequelabs/axe-core-npm/issues/727
+    const timeoutPromise = new Promise((resolve, reject) => {
+      setTimeout(() => {
+        reject();
+      }, 1000);
+    });
+    const evaluatePromise = frame.evaluate(pageIsLoaded);
+    const readyState = await Promise.race([timeoutPromise, evaluatePromise]);
+    assert(readyState);
   } catch {
-    /* ignore */
-  }
-  if (!pageReady) {
     throw new Error('Page/Frame is not ready');
   }
 }