+++ packages/react/after.ts
@@ -1,22 +1,26 @@
 let restoreFunctions: Function[] = [];
 
-function after(host: React.Component, name: string, cb: Function): void {
+export default function after(
+  host: React.Component,
+  name: string,
+  cb: Function
+): void {
   const originalFn: Function = host[name];
   let restoreFn: () => void;
 
   if (originalFn) {
-    host[name] = function(...args): void {
+    host[name] = function (...args): void {
       originalFn.apply(this, args);
       cb(host);
     };
-    restoreFn = function(): void {
+    restoreFn = function (): void {
       host[name] = originalFn;
     };
   } else {
-    host[name] = function(): void {
+    host[name] = function (): void {
       cb(host);
     };
-    restoreFn = function(): void {
+    restoreFn = function (): void {
       delete host[name];
     };
   }
@@ -24,9 +28,7 @@ function after(host: React.Component, name: string, cb: Function): void {
   restoreFunctions.push(restoreFn);
 }
 
-after.restorePatchedMethods = function(): void {
+after.restorePatchedMethods = function (): void {
   restoreFunctions.forEach(restoreFn => restoreFn());
   restoreFunctions = [];
 };
-
-export = after;