+++ packages/reporter-earl/esmTest.mjs
@@ -1,0 +1,5 @@
+import defaultExport from './dist/axeReporterEarl.mjs';
+import assert from 'assert';
+
+const exportIsFunction = typeof defaultExport === 'function';
+assert(exportIsFunction, 'export is not a function');