+++ CHANGELOG.md
@@ -3,36 +3,69 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# [4.6.0](https://github.com/dequelabs/axe-core-npm/compare/v4.5.2...v4.6.0) (2023-01-19)
-
+## [4.7.3](https://github.com/dequelabs/axe-core-npm/compare/v4.7.2...v4.7.3) (2023-06-09)
 
 ### Bug Fixes
 
-* **playwright:** update playwright to support updated types ([#635](https://github.com/dequelabs/axe-core-npm/issues/635)) ([1339ec7](https://github.com/dequelabs/axe-core-npm/commit/1339ec7199fd654f6116ecaf824274a163cf3a73))
-* use `SerialContextObject` type instead of `ContextObject` ([#632](https://github.com/dequelabs/axe-core-npm/issues/632)) ([630b545](https://github.com/dequelabs/axe-core-npm/commit/630b545469b8e973e2aa9c7d879f9c120ce8a4d9))
+- **webdriverjs:** skip unloaded iframes ([#743](https://github.com/dequelabs/axe-core-npm/issues/743)) ([9cb50eb](https://github.com/dequelabs/axe-core-npm/commit/9cb50ebb6b138cb0aa6b785bc7c826c4b9a21086))
 
+## [4.7.2](https://github.com/dequelabs/axe-core-npm/compare/v4.7.1...v4.7.2) (2023-06-01)
+
+### Bug Fixes
+
+- **packages/cli:** revert ESM ([#730](https://github.com/dequelabs/axe-core-npm/issues/730)) ([4e1fb95](https://github.com/dequelabs/axe-core-npm/commit/4e1fb958fa251db7eea3433f91f2a2f3adbda1e8))
+- **playwright:** Enable usage with `@playwright/test` ([#731](https://github.com/dequelabs/axe-core-npm/issues/731)) ([0d0c9d1](https://github.com/dequelabs/axe-core-npm/commit/0d0c9d1b75051b26f0bb745abc5c8974a9fa4d54))
+
+## [4.7.1](https://github.com/dequelabs/axe-core-npm/compare/v4.7.0...v4.7.1) (2023-05-02)
+
+### Bug Fixes
+
+- proper imports in esm context ([#718](https://github.com/dequelabs/axe-core-npm/issues/718)) ([f2a41bc](https://github.com/dequelabs/axe-core-npm/commit/f2a41bcaf33f5000fa1cd08f1a3e184b5d6360be))
+- **react:** include .mjs file in publish ([#722](https://github.com/dequelabs/axe-core-npm/issues/722)) ([a7bf92f](https://github.com/dequelabs/axe-core-npm/commit/a7bf92fd9856b7be0fea4f3c8bd38a79a3c9fde3))
+
+# [4.7.0](https://github.com/dequelabs/axe-core-npm/compare/v4.6.1...v4.7.0) (2023-04-28)
+
+### Bug Fixes
+
+- **cli:** do not inject script tags due to pages not allowing script tags ([#710](https://github.com/dequelabs/axe-core-npm/issues/710)) ([2e0fdb4](https://github.com/dequelabs/axe-core-npm/commit/2e0fdb4a4661040b9b8dbb91bed1a96142e4fe0c))
 
 ### Features
 
-* support shadow dom selectors ([#619](https://github.com/dequelabs/axe-core-npm/issues/619)) ([6065b56](https://github.com/dequelabs/axe-core-npm/commit/6065b562adf289341033a7233426ba3dcfb8a24b))
-* Update axe-core to v4.6.1 ([#633](https://github.com/dequelabs/axe-core-npm/issues/633)) ([8bab5b2](https://github.com/dequelabs/axe-core-npm/commit/8bab5b23b3653a6a0603f10d889e22418a4fff3e))
+- enable esm import ([#713](https://github.com/dequelabs/axe-core-npm/issues/713)) ([076bfa9](https://github.com/dequelabs/axe-core-npm/commit/076bfa99f27f416b4eaff6d24cc23e8573fc72ee))
+- support WDIO v8 ([#708](https://github.com/dequelabs/axe-core-npm/issues/708)) ([1e40ec0](https://github.com/dequelabs/axe-core-npm/commit/1e40ec006844a9b5f0e368b5ca8014e215c75372))
+- Update axe-core to v4.7.0 ([#709](https://github.com/dequelabs/axe-core-npm/issues/709)) ([d418d09](https://github.com/dequelabs/axe-core-npm/commit/d418d093b4c462d5d60e0f7ad3d5147bcaf52aa2))
 
-
-
-
-
-## [4.5.2](https://github.com/dequelabs/axe-core-npm/compare/v4.5.1...v4.5.2) (2022-11-15)
-
+## [4.6.1](https://github.com/dequelabs/axe-core-npm/compare/v4.6.0...v4.6.1) (2023-03-28)
 
 ### Bug Fixes
 
-* **cli:** include all functions in types file ([#603](https://github.com/dequelabs/axe-core-npm/issues/603)) ([30042ef](https://github.com/dequelabs/axe-core-npm/commit/30042ef3acc0cab7f67ddcf284df089f243f53e6))
-* **cli:** use correct driver typings ([#604](https://github.com/dequelabs/axe-core-npm/issues/604)) ([d33bbdf](https://github.com/dequelabs/axe-core-npm/commit/d33bbdf2b3d05c3ba0106d717ba1c8d706c5e05c))
-* update to use axe-core 4.5.2 ([#607](https://github.com/dequelabs/axe-core-npm/issues/607)) ([61312ff](https://github.com/dequelabs/axe-core-npm/commit/61312fff2e9ee4e0f218a29a1608f69b45206fad))
+- **cli:** increase --timeout default to 90 seconds (to match --help) ([#683](https://github.com/dequelabs/axe-core-npm/issues/683)) ([d11481b](https://github.com/dequelabs/axe-core-npm/commit/d11481b8c74598617260586fe14c8ceba32c6792))
+- Update axe-core to v4.6.3 ([#698](https://github.com/dequelabs/axe-core-npm/issues/698)) ([a7d53a4](https://github.com/dequelabs/axe-core-npm/commit/a7d53a4876237c74bb8dc2836b703cdaf2f7ace5))
+- work with extremely large results ([#667](https://github.com/dequelabs/axe-core-npm/issues/667)) ([395d5fc](https://github.com/dequelabs/axe-core-npm/commit/395d5fc165c97f0c297176eb69eb58d2510c938f))
 
+### Features
 
+- **webdriverio:** deprecate @wdio/sync support ([#677](https://github.com/dequelabs/axe-core-npm/issues/677)) ([6121e55](https://github.com/dequelabs/axe-core-npm/commit/6121e55e000b0b8b5b93c92b8530cfa1bd505da8))
 
+# [4.6.0](https://github.com/dequelabs/axe-core-npm/compare/v4.5.2...v4.6.0) (2023-01-19)
 
+### Bug Fixes
+
+- **playwright:** update playwright to support updated types ([#635](https://github.com/dequelabs/axe-core-npm/issues/635)) ([1339ec7](https://github.com/dequelabs/axe-core-npm/commit/1339ec7199fd654f6116ecaf824274a163cf3a73))
+- use `SerialContextObject` type instead of `ContextObject` ([#632](https://github.com/dequelabs/axe-core-npm/issues/632)) ([630b545](https://github.com/dequelabs/axe-core-npm/commit/630b545469b8e973e2aa9c7d879f9c120ce8a4d9))
+
+### Features
+
+- support shadow dom selectors ([#619](https://github.com/dequelabs/axe-core-npm/issues/619)) ([6065b56](https://github.com/dequelabs/axe-core-npm/commit/6065b562adf289341033a7233426ba3dcfb8a24b))
+- Update axe-core to v4.6.1 ([#633](https://github.com/dequelabs/axe-core-npm/issues/633)) ([8bab5b2](https://github.com/dequelabs/axe-core-npm/commit/8bab5b23b3653a6a0603f10d889e22418a4fff3e))
+
+## [4.5.2](https://github.com/dequelabs/axe-core-npm/compare/v4.5.1...v4.5.2) (2022-11-15)
+
+### Bug Fixes
+
+- **cli:** include all functions in types file ([#603](https://github.com/dequelabs/axe-core-npm/issues/603)) ([30042ef](https://github.com/dequelabs/axe-core-npm/commit/30042ef3acc0cab7f67ddcf284df089f243f53e6))
+- **cli:** use correct driver typings ([#604](https://github.com/dequelabs/axe-core-npm/issues/604)) ([d33bbdf](https://github.com/dequelabs/axe-core-npm/commit/d33bbdf2b3d05c3ba0106d717ba1c8d706c5e05c))
+- update to use axe-core 4.5.2 ([#607](https://github.com/dequelabs/axe-core-npm/issues/607)) ([61312ff](https://github.com/dequelabs/axe-core-npm/commit/61312fff2e9ee4e0f218a29a1608f69b45206fad))
 
 ## [4.5.1](https://github.com/dequelabs/axe-core-npm/compare/v4.5.0...v4.5.1) (2022-11-03)