+++ packages/webdriverio/CHANGELOG.md
@@ -3,33 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# [4.6.0](https://github.com/dequelabs/axe-core-npm/compare/v4.5.2...v4.6.0) (2023-01-19)
+## [4.7.3](https://github.com/dequelabs/axe-core-npm/compare/v4.7.2...v4.7.3) (2023-06-09)
 
+**Note:** Version bump only for package @axe-core/webdriverio
+
+## [4.7.2](https://github.com/dequelabs/axe-core-npm/compare/v4.7.1...v4.7.2) (2023-06-01)
+
+**Note:** Version bump only for package @axe-core/webdriverio
+
+## [4.7.1](https://github.com/dequelabs/axe-core-npm/compare/v4.7.0...v4.7.1) (2023-05-02)
 
 ### Bug Fixes
 
-* use `SerialContextObject` type instead of `ContextObject` ([#632](https://github.com/dequelabs/axe-core-npm/issues/632)) ([630b545](https://github.com/dequelabs/axe-core-npm/commit/630b545469b8e973e2aa9c7d879f9c120ce8a4d9))
+- proper imports in esm context ([#718](https://github.com/dequelabs/axe-core-npm/issues/718)) ([f2a41bc](https://github.com/dequelabs/axe-core-npm/commit/f2a41bcaf33f5000fa1cd08f1a3e184b5d6360be))
 
+# [4.7.0](https://github.com/dequelabs/axe-core-npm/compare/v4.6.1...v4.7.0) (2023-04-28)
 
 ### Features
 
-* support shadow dom selectors ([#619](https://github.com/dequelabs/axe-core-npm/issues/619)) ([6065b56](https://github.com/dequelabs/axe-core-npm/commit/6065b562adf289341033a7233426ba3dcfb8a24b))
-* Update axe-core to v4.6.1 ([#633](https://github.com/dequelabs/axe-core-npm/issues/633)) ([8bab5b2](https://github.com/dequelabs/axe-core-npm/commit/8bab5b23b3653a6a0603f10d889e22418a4fff3e))
+- enable esm import ([#713](https://github.com/dequelabs/axe-core-npm/issues/713)) ([076bfa9](https://github.com/dequelabs/axe-core-npm/commit/076bfa99f27f416b4eaff6d24cc23e8573fc72ee))
+- support WDIO v8 ([#708](https://github.com/dequelabs/axe-core-npm/issues/708)) ([1e40ec0](https://github.com/dequelabs/axe-core-npm/commit/1e40ec006844a9b5f0e368b5ca8014e215c75372))
+- Update axe-core to v4.7.0 ([#709](https://github.com/dequelabs/axe-core-npm/issues/709)) ([d418d09](https://github.com/dequelabs/axe-core-npm/commit/d418d093b4c462d5d60e0f7ad3d5147bcaf52aa2))
 
-
-
-
-
-## [4.5.2](https://github.com/dequelabs/axe-core-npm/compare/v4.5.1...v4.5.2) (2022-11-15)
-
+## [4.6.1](https://github.com/dequelabs/axe-core-npm/compare/v4.6.0...v4.6.1) (2023-03-28)
 
 ### Bug Fixes
 
-* update to use axe-core 4.5.2 ([#607](https://github.com/dequelabs/axe-core-npm/issues/607)) ([61312ff](https://github.com/dequelabs/axe-core-npm/commit/61312fff2e9ee4e0f218a29a1608f69b45206fad))
+- Update axe-core to v4.6.3 ([#698](https://github.com/dequelabs/axe-core-npm/issues/698)) ([a7d53a4](https://github.com/dequelabs/axe-core-npm/commit/a7d53a4876237c74bb8dc2836b703cdaf2f7ace5))
+- work with extremely large results ([#667](https://github.com/dequelabs/axe-core-npm/issues/667)) ([395d5fc](https://github.com/dequelabs/axe-core-npm/commit/395d5fc165c97f0c297176eb69eb58d2510c938f))
 
+### Features
 
+- **webdriverio:** deprecate @wdio/sync support ([#677](https://github.com/dequelabs/axe-core-npm/issues/677)) ([6121e55](https://github.com/dequelabs/axe-core-npm/commit/6121e55e000b0b8b5b93c92b8530cfa1bd505da8))
 
+# [4.6.0](https://github.com/dequelabs/axe-core-npm/compare/v4.5.2...v4.6.0) (2023-01-19)
 
+### Bug Fixes
+
+- use `SerialContextObject` type instead of `ContextObject` ([#632](https://github.com/dequelabs/axe-core-npm/issues/632)) ([630b545](https://github.com/dequelabs/axe-core-npm/commit/630b545469b8e973e2aa9c7d879f9c120ce8a4d9))
+
+### Features
+
+- support shadow dom selectors ([#619](https://github.com/dequelabs/axe-core-npm/issues/619)) ([6065b56](https://github.com/dequelabs/axe-core-npm/commit/6065b562adf289341033a7233426ba3dcfb8a24b))
+- Update axe-core to v4.6.1 ([#633](https://github.com/dequelabs/axe-core-npm/issues/633)) ([8bab5b2](https://github.com/dequelabs/axe-core-npm/commit/8bab5b23b3653a6a0603f10d889e22418a4fff3e))
+
+## [4.5.2](https://github.com/dequelabs/axe-core-npm/compare/v4.5.1...v4.5.2) (2022-11-15)
+
+### Bug Fixes
+
+- update to use axe-core 4.5.2 ([#607](https://github.com/dequelabs/axe-core-npm/issues/607)) ([61312ff](https://github.com/dequelabs/axe-core-npm/commit/61312fff2e9ee4e0f218a29a1608f69b45206fad))
 
 ## [4.5.1](https://github.com/dequelabs/axe-core-npm/compare/v4.5.0...v4.5.1) (2022-11-03)